fix: Auth Provider Added
